chore: completes framework and tests

+++ lesson_04/expression/expression_app/src/main/java/com/codedifferently/lesson4/ExpressionCalculator.java
@@ -1,8 +1,15 @@
 package com.codedifferently.lesson4;
 
 public class ExpressionCalculator {
-
-  public Double calculate(int a, int b, int c, int d, int e) {
-    return 0.0;
+  /**
+   * Returns the calculation of ((a + b) * c) / d^e
+   *
+   * @param a
+   * @param b
+   * @return
+   */
+  public double calculate(double a, double b, double c, double d, double e) {
+      // Implement your code here to return the correct value.
+    return 0d;
   }
 }
